Stupid error in this code was fixed
The big bug is the Serial.println() in the readXbee function that is
antiquated and no longer needing to be used. In fact, this is what
caused the Zigbee to exit from Command Mode when I invoked "+++", then
made the Xbee module receive a data packet.
Will try to make things compatible with another instance of SoftSerial
for the Xbee.